+++ src/main/java/za/ca/cput/adp3capstone/repository/TravelPlanRepository.java
@@ -1,4 +1,4 @@
-package za.ca.cput.adp3capstone.repisitory;
+package za.ca.cput.adp3capstone.repository;
 
 
 import za.ca.cput.adp3capstone.domain.TravelPlan;
