+++ src/test/java/za/ca/cput/adp3capstone/factory/ItineraryFactoryTest.java
@@ -46,6 +46,5 @@ class ItineraryFactoryTest {
     @Order(4)
     @Disabled
     public void testNotImplementedYet() {
-        // Todo
     }
 }
